change name to table-of-contents-seo to prevent regression with prior toc block
